refactor separate Pin class
